requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aenum==3.1.15
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
bddl==3.5.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blobfile==3.0.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.6.0
cmake==3.30.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.31
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.29.2
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.23
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
egl_probe==1.0.2
einops==0.8.0
etils==1.3.0
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.13
fasteners==0.15
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
free-mujoco-py==2.1.6
freetype-py==2.5.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.65.2
gym==0.17.3
gym-notices==0.0.8
h5py==3.11.0
huggingface-hub==0.24.5
Hydra==2.5
hydra-core==1.3.2
idna==3.7
igibson==2.2.1
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
jupytext==1.16.4
keras==2.13.1
kiwisolver==1.4.5
labmaze==1.0.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
lpips==0.1.4
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
patchelf==0.17.2.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
piq==0.6.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
proglog==0.1.10
progressbar==2.5
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py360convert==0.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet-svl==3.1.6.4
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyglet==1.5.0
Pygments==2.18.0
pynput==1.7.7
PyOpenGL==3.1.7
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
robodesk==1.0.0
robomimic==0.2.0
robosuite==1.4.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stable-baselines3==1.4.0
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.21.0
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
timm==1.0.8
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torch-fidelity==0.3.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
transforms3d==0.4.2
trimesh==4.4.9
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.5
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
yarg==0.1.9
zipp==3.19.2
